+++ ui/src/app/app.module.ts
@@ -28,6 +28,7 @@ import {FetchModule} from './services/fetch.module';
 import {TextModalComponent} from 'app/components/text-modal';
 import {NavBarComponent} from 'app/pages/signed-in/nav-bar';
 import {FooterComponent} from './components/footer';
+import {ZendeskWidgetComponent} from './components/zendesk-widget';
 
 
 
@@ -59,6 +60,7 @@ import {FooterComponent} from './components/footer';
     TextModalComponent,
     WorkspaceNavBarComponent,
     WorkspaceWrapperComponent,
+    ZendeskWidgetComponent
   ],
   providers: [
     SignInService,
