+++ ui/src/app/utils/zendesk.ts
@@ -1,4 +1,3 @@
-import {ElementRef} from '@angular/core';
 import {environment} from 'environments/environment';
 import {ZendeskEnv} from 'environments/environment-type';
 
@@ -27,6 +26,10 @@ const zendeskConfigs = {
     baseUrl: 'https://aousupporthelp1580753096.zendesk.com/hc',
     widgetKey: 'df0a2e39-f8a8-482b-baf5-af82e14d38f9'
   }
+};
+
+export const zendeskWidgetKey = () => {
+  return zendeskConfigs[environment.zendeskEnv].widgetKey;
 };
 
 /**
@@ -84,50 +87,6 @@ export const supportUrls: ZendeskUrls = ((env) => {
 // ideally this initialization would be done by including <script> tags in
 // the component template html file.  However, Angular does not allow this.
 // TODO: investigate whether this will be possible after we migrate to React
-
-export function initializeZendeskWidget(elementRef: ElementRef): void {
-  const accessKey = zendeskConfigs[environment.zendeskEnv].widgetKey;
-
-  // This external script loads the Zendesk web widget, connected to our
-  // production Zendesk account.
-  const s = document.createElement('script');
-  s.type = 'text/javascript';
-  s.id = 'ze-snippet';
-  s.src = 'https://static.zdassets.com/ekr/snippet.js?key=' + accessKey;
-  elementRef.nativeElement.appendChild(s);
-
-  // This data configures the Zendesk web widget with settings to show only
-  // the "contact us" form. See https://developer.zendesk.com/embeddables/docs/widget/
-  // for API docs.
-  window['zESettings'] = {
-    webWidget: {
-      chat: {
-        suppress: true,
-      },
-      color: {
-        // This is an AoU dark purple color.
-        theme: '#262262'
-      },
-      contactForm: {
-        attachments: true,
-        subject: false,
-        // We include a tag indicating that this support request was filed via the
-        // Researcher Workbench. This helps distinguish from tickets filed via other
-        // AoU sub-products, e.g. Research Hub or Data Browser.
-        tags: ['research_workbench'],
-        title: {
-          '*': 'Help Desk',
-        },
-      },
-      helpCenter: {
-        suppress: true,
-      },
-      talk: {
-        suppress: true,
-      },
-    }
-  };
-}
 
 export function openZendeskWidget(
   givenName: string, familyName: string, aouEmailAddress: string,
